web_text_classify/requirements.txt
absl-py==0.11.0
aiohttp==3.7.3
albumentations==0.5.2
alembic==1.4.3
aliyun-python-sdk-core==2.13.30
aliyun-python-sdk-core-v3==2.13.11
aliyun-python-sdk-kms==2.13.0
asgiref==3.3.1
astor==0.8.1
astunparse==1.6.3
async-generator==1.10
async-timeout==3.0.1
attrs==20.3.0
backcall==0.2.0
bleach==3.2.1
cached-property==1.5.2
cachetools==4.1.1
certifi==2020.11.8
certipy==0.1.3
cffi==1.14.4
chardet==3.0.4
colorama==0.4.4
contextlib2==0.6.0.post1
crcmod==1.7
cryptography==3.2.1
cycler==0.10.0
dataclasses==0.6
decorator==4.4.2
defusedxml==0.6.0
Django==3.1.4
djangorestframework==3.12.2
entrypoints==0.3
featurize-serverext @ file:///tmp/serverext
future==0.18.2
gast==0.3.3
gensim==3.8.3
google-auth==1.23.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
GPUtil==1.4.0
grpcio==1.34.0
gunicorn==20.0.4
h5py==2.10.0
hyperopt==0.1.2
idna==2.10
imageio==2.9.0
imgaug==0.4.0
importlib-metadata==3.1.1
ipykernel==5.3.4
ipython==7.19.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.17.2
Jinja2==2.11.2
jmespath==0.10.0
joblib==0.17.0
json-tricks==3.15.5
json5==0.9.5
jsonschema==3.2.0
jupyter-client==6.1.7
jupyter-core==4.7.0
jupyter-server-proxy @ git+https://github.com/louis-she/jupyter-server-proxy.git@4959711d122f10b6c0442dcc9c7362fb2bb98db3
jupyter-telemetry==0.1.0
jupyterhub==1.1.0
jupyterlab==2.1.0
jupyterlab-pygments==0.1.2
jupyterlab-server==1.2.0
kaggle==1.5.10
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
lightgbm==3.1.0.99
Mako==1.1.3
Markdown==3.3.3
MarkupSafe==1.1.1
matplotlib==3.3.3
mistune==0.8.4
mock==4.0.2
multidict==5.1.0
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.3
netifaces==0.10.9
networkx==2.5
notebook==6.0.3
numpy==1.18.5
oauthlib==3.1.0
opencv-python==4.4.0.46
opencv-python-headless==4.4.0.46
opt-einsum==3.3.0
oss2==2.13.1
packaging==20.7
pamela==1.0.0
pandas==1.1.4
pandocfilters==1.4.3
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.0.1
pkginfo==1.6.1
plotly==4.13.0
portpicker==1.3.1
prometheus-client==0.9.0
prompt-toolkit==3.0.8
protobuf==3.14.0
psutil==5.7.0
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodome==3.9.9
Pygments==2.7.2
pymongo==3.11.2
pyOpenSSL==20.0.0
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
python-editor==1.0.4
python-json-logger==2.0.1
python-slugify==4.0.1
PythonWebHDFS==0.2.3
pytz==2020.4
PyWavelets==1.1.1
PyYAML==5.3.1
pyzmq==20.0.0
requests==2.25.0
requests-oauthlib==1.3.0
retrying==1.3.3
rsa==4.6
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
schema==0.7.3
scikit-image==0.17.2
scikit-learn==0.21.3
scipy==1.5.4
Send2Trash==1.5.0
Shapely==1.7.1
simpervisor==0.3
simplejson==3.17.2
six==1.15.0
sklearn==0.0
smart-open==4.0.1
SQLAlchemy==1.3.20
sqlparse==0.4.1
tensorboard==2.4.0
tensorboard-plugin-wit==1.7.0
tensorflow-estimator==2.3.0
tensorflow-gpu==2.3.1
termcolor==1.1.0
terminado==0.9.1
testpath==0.4.4
text-unidecode==1.3
threadpoolctl==2.1.0
tifffile==2020.11.26
torch==1.7.0
torchaudio==0.7.0
torchtext==0.8.0
torchvision==0.8.1
tornado==6.1
tqdm==4.54.0
traitlets==5.0.5
typing-extensions==3.7.4.3
urllib3==1.26.2
wcwidth==0.2.5
webencodings==0.5.1
websockets==8.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
yarl==1.6.3
zipp==3.4.0
keras_bert==0.86.0
tensorflow_addons==0.11.2
